txmod-client/requirements.txt
astroid==2.3.3
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
certifi==2020.4.5.1
chardet==3.0.4
cytoolz==0.10.1
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
googleapis-common-protos==1.51.0
grpcio==1.28.1
grpcio-status==1.28.1
grpcio-tools==1.28.1
hexbytes==0.2.0
idna==2.9
importlib==1.0.4
importlib-metadata==1.6.0
ipfshttpclient==0.4.12
isort==4.3.21
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.6
mccabe==0.6.1
multiaddr==0.0.9
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.11.3
pycryptodome==3.9.7
pylint==2.4.4
pyrsistent==0.16.0
requests==2.23.0
rlp==1.2.0
six==1.14.0
termcolor==1.1.0
toolz==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
varint==1.0.2
web3==5.7.0
websockets==8.1
wrapt==1.11.2
zipp==3.1.0
